requirements.txt
asgiref==3.2.3
Django==3.0.3
django-modeltranslation==0.14.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
